Move Vector into utilities package
